+++ test/dotnet-watch.Tests/Browser/BrowserConnectorTests.cs
@@ -18,6 +18,6 @@ public class BrowserConnectorTests
     [InlineData("https://localhost:1000/x/y?z=u", "https://localhost:1234/a?b=c", "https://localhost:1000/x/y?z=u")]
     public void GetLaunchUrl(string? profileLaunchUrl, string outputLaunchUrl, string expected)
     {
-        Assert.Equal(expected, BrowserConnector.GetLaunchUrl(profileLaunchUrl, outputLaunchUrl));
+        Assert.Equal(expected, BrowserRefreshServerFactory.GetLaunchUrl(profileLaunchUrl, outputLaunchUrl));
     }
 }
